Set the ownership of existing directories too.
When mounting a data-volume as a subdirectory
inside the users home directory and explicitly
specifying it as the upload directory, this
directory will be chowned to $uid:users.

In the example below the user can upload files
directly to the root folder of the specified
volume 'my_data_vol' as expected.

docker-compose.yml:

    ...

    volumes:
        - my_data_vol:/home/miguel/www

    command: miguel:secret:::www

    ...