[PATCH] AM-76 Only use the PAM script for session actions
The PAM script only handles the session type and returns true in all other
cases.   It was being symlinked to /usr/share/libpam-script/pam_script_auth,
which caused authentication attempts for any valid user to be accepted, with
any password.  It was also linked to pam_script_acct, pam_script_auth and
pam_script_passwd which are similarly inappropriate.

Stop creating these symlinks to prevent this.

Signed-off-by: Yasser Saleemi <yassersaleemi@gmail.com>